Move spinner to ui mod.
